refactor: mapper 폴더명 수정
