+++ src/main/java/com/recipe/app/src/common/application/AppVersionService.java
@@ -1,7 +1,7 @@
 package com.recipe.app.src.common.application;
 
 import com.recipe.app.src.common.domain.AppVersion;
-import com.recipe.app.src.common.mapper.AppVersionRepositiory;
+import com.recipe.app.src.common.infra.AppVersionRepositiory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
