+++ src/main/java/com/recipe/app/src/common/application/BadWordService.java
@@ -2,7 +2,7 @@ package com.recipe.app.src.common.application;
 
 import com.recipe.app.src.common.domain.BadWords;
 import com.recipe.app.src.common.exception.BadWordException;
-import com.recipe.app.src.common.mapper.BadWordsRepository;
+import com.recipe.app.src.common.infra.BadWordsRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
