+++ src/main/java/com/recipe/app/src/common/application/DialogService.java
@@ -1,14 +1,10 @@
 package com.recipe.app.src.common.application;
 
-import com.recipe.app.common.exception.BaseException;
-import com.recipe.app.src.common.application.dto.DialogDto;
 import com.recipe.app.src.common.domain.Dialog;
 import com.recipe.app.src.common.exception.NotFoundNoticeException;
-import com.recipe.app.src.common.mapper.DialogRepository;
+import com.recipe.app.src.common.infra.DialogRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
-import static com.recipe.app.common.response.BaseResponseStatus.NOT_FOUND_NOTICE;
 
 @Service
 public class DialogService {
